Move Tabular to its own file. Handle GTC format files
