+++ fixtures/input.json
@@ -1,0 +1,5 @@
+{
+  "file_relationships":[
+      "http://stemcellcommons.org/sites/default/files/gct/7283_GPL1261.gct"
+   ]
+}
